reduce diff in tensor_ops.hpp
